Remove average score from TODO
This is calculated by taking the sum of score field and dividing it by
the amount of plays.

What's strange about this is that the the amount of plays goes up when
you exit to the title screen during a run, but the sum of score field does
not go up. So the 'average' score reported in the player profile is not
actually correct.