Adds some fields (#1)
* Add area death count

* Remove deadliest level from TODO

Deadliest level is handled by the 'Area - Death Count' fields.

* Add character death count

* Add player character selection

* Add sum of score field

* Remove average score from TODO

This is calculated by taking the sum of score field and dividing it by
the amount of plays.

What's strange about this is that the the amount of plays goes up when
you exit to the title screen during a run, but the sum of score field does
not go up. So the 'average' score reported in the player profile is not
actually correct.

* Add sum of time field

Belongs to the player profile but it's at a totally different address.

I decided to still place it in the profile section. :/

* Remove average time from TODO

This is calculated by taking the sum of time and dividing it by the play
count.

The average time stat suffers from the same problem described in commit
e07ec436bf95.

* Add best time field

Removes the TODO for the time field and for the player profile since
everything on that page has been accounted for.

* Add area 7-[1234] death count

* Introduce 64 bit types