Large refactor and fixes for 1.17
This change allows you to pass multiple search directories
for finding packs. It keeps the overrides directory for now as the default.

This change also allows files to be anywhere in a search directory. This allows
for simple "Pack" dirs that only have a single file without worrying about
directory layout.

Updates for 1.17:
    Updates known asset list to use DDS extension on textures
    Added an exlusion set of dds images that don't convert cleanly
    Adds code for converting back to dds from png (Thanks Cloppershy!)
    Adds new strings08.str to known assets