Unify integer type implementations
Adding new integer types was annoying and error prone.

It's unfortunate that some types are classes and some are instances, but
that was already the case before this commit.

I don't really think it's a huge issue. In the future this could be
solved by making every reused type an instance and giving the classes a
different suffix, i.e. 'TypeDef' instead of 'Type'.